+++ mono_repo/test/generate_test.dart
@@ -652,6 +652,29 @@ ${_writeScriptOutput(false)}''',
     },
   );
 
+  test('no unused `os` warning when top-level `os` is omitted', () async {
+    await d.dir('pkg_a', [
+      d.file(monoPkgFileName, r'''
+stages:
+  - unit_test:
+    - test:
+      sdk:
+        - dev
+      os:
+        - windows
+'''),
+      d.file('pubspec.yaml', '''
+name: pkg_a
+      '''),
+    ]).create();
+
+    testGenerateConfig(
+      printMatcher: '''
+package:pkg_a
+${_writeScriptOutput(false)}''',
+    );
+  });
+
   test('test_with_coverage', () async {
     await d.dir('pkg_a', [
       d.file(monoPkgFileName, r'''